split up server.js
