+++ backend/macros.ts
@@ -2,13 +2,14 @@
  * This file is part of Search NEU and licensed under AGPL3.
  * See the license file in the root folder for details.
  */
+import {Request} from 'express';
+import path from "path";
+import fs from "fs-extra";
+import Rollbar, { MaybeError } from "rollbar";
+import Amplitude from "amplitude";
 
-import path from 'path';
-import fs from 'fs-extra';
-import Rollbar, { MaybeError } from 'rollbar';
-import Amplitude from 'amplitude';
-
-import commonMacros from '../common/abstractMacros';
+import commonMacros from "../common/abstractMacros";
+import moment from "moment";
 
 const amplitude = new Amplitude(commonMacros.amplitudeToken);
 
@@ -25,16 +26,19 @@ const originalCwd: string = process.cwd();
 let oldcwd: string;
 while (1) {
   try {
-    fs.statSync('package.json');
+    fs.statSync("package.json");
   } catch (e) {
     oldcwd = process.cwd();
     //cd .. until in the same dir as package.json, the root of the project
-    process.chdir('..');
+    process.chdir("..");
 
     // Prevent an infinate loop: If we keep cd'ing upward and we hit the root dir and still haven't found
     // a package.json, just return to the original directory and break out of this loop.
     if (oldcwd === process.cwd()) {
-      commonMacros.warn("Can't find directory with package.json, returning to", originalCwd);
+      commonMacros.warn(
+        "Can't find directory with package.json, returning to",
+        originalCwd
+      );
       process.chdir(originalCwd);
       break;
     }
@@ -44,18 +48,19 @@ while (1) {
   break;
 }
 
-type EnvKeys = 'elasticURL'
-| 'dbName'
-| 'dbHost'
-// Secrets:
-| 'dbUsername'
-| 'dbPassword'
-| 'rollbarPostServerItemToken'
-| 'fbToken'
-| 'fbVerifyToken'
-| 'fbAppSecret'
-// Only for dev:
-| 'fbMessengerId';
+type EnvKeys =
+  | "elasticURL"
+  | "dbName"
+  | "dbHost"
+  // Secrets:
+  | "dbUsername"
+  | "dbPassword"
+  | "rollbarPostServerItemToken"
+  | "fbToken"
+  | "fbVerifyToken"
+  | "fbAppSecret"
+  // Only for dev:
+  | "fbMessengerId";
 
 type EnvVars = Partial<Record<EnvKeys, string>>;
 
@@ -66,29 +71,29 @@ type EnvVars = Partial<Record<EnvKeys, string>>;
 let envVariables: EnvVars = null;
 
 class Macros extends commonMacros {
-// Version of the schema for the data. Any changes in this schema will effect the data saved in the dev_data folder
+  // Version of the schema for the data. Any changes in this schema will effect the data saved in the dev_data folder
   // and the data saved in the term dumps in the public folder and the search indexes in the public folder.
   // Increment this number every time there is a breaking change in the schema.
   // This will cause the data to be saved in a different folder in the public data folder.
   // The first schema change is here: https://github.com/ryanhugh/searchneu/pull/48
   static schemaVersion = 2;
 
-  static PUBLIC_DIR = path.join('public', 'data', `v${Macros.schemaVersion}`);
+  static PUBLIC_DIR = path.join("public", "data", `v${Macros.schemaVersion}`);
 
-  static DEV_DATA_DIR = path.join('dev_data', `v${Macros.schemaVersion}`);
+  static DEV_DATA_DIR = path.join("dev_data", `v${Macros.schemaVersion}`);
 
   // Folder of the raw html cache for the requests.
-  static REQUESTS_CACHE_DIR = 'requests';
+  static REQUESTS_CACHE_DIR = "requests";
 
   // For iterating over every letter in a couple different places in the code.
-  static ALPHABET = 'maqwertyuiopsdfghjklzxcvbn';
+  static ALPHABET = "maqwertyuiopsdfghjklzxcvbn";
 
   static getAllEnvVariables(): EnvVars {
     if (envVariables) {
       return envVariables;
     }
 
-    let configFileName = '/etc/searchneu/config.json';
+    let configFileName = "/etc/searchneu/config.json";
 
     // Yes, this is syncronous instead of the normal Node.js async style
     // But keeping it sync helps simplify other parts of the code
@@ -98,7 +103,7 @@ class Macros extends commonMacros {
 
     // Also check /mnt/c/etc... in case we are running inside WSL.
     if (!exists) {
-      configFileName = '/mnt/c/etc/searchneu/config.json';
+      configFileName = "/mnt/c/etc/searchneu/config.json";
       exists = fs.existsSync(configFileName);
     }
 
@@ -111,6 +116,37 @@ class Macros extends commonMacros {
     envVariables = Object.assign(envVariables, process.env);
 
     return envVariables;
+  }
+
+  // Gets the current time, just used for loggin
+  static getTime() {
+    return moment().format("hh:mm:ss a");
+  }
+
+  // Prefer the headers if they are present so we get the real ip instead of localhost (nginx) or a cloudflare IP
+  static getIpPath(req: Request) {
+    const output = [];
+
+    const realIpHeader = req.headers["x-real-ip"];
+    if (realIpHeader) {
+      output.push("Real:");
+      output.push(realIpHeader);
+      output.push(" ");
+    }
+
+    const forwardedForHeader = req.headers["x-forwarded-for"];
+    if (forwardedForHeader) {
+      output.push("ForwardedFor:");
+      output.push(forwardedForHeader);
+      output.push(" ");
+    }
+
+    if (req.connection.remoteAddress !== "127.0.0.1") {
+      output.push("remoteIp: ");
+      output.push(req.connection.remoteAddress);
+    }
+
+    return output.join("");
   }
 
   static getEnvVariable(name: EnvKeys): string {
@@ -131,11 +167,11 @@ class Macros extends commonMacros {
     };
 
     return amplitude.track(data).catch((error) => {
-      Macros.warn('error Logging amplitude event failed:', error);
+      Macros.warn("error Logging amplitude event failed:", error);
     });
   }
   static getRollbar(...args: any) {
-    const rollbarKey = Macros.getEnvVariable('rollbarPostServerItemToken');
+    const rollbarKey = Macros.getEnvVariable("rollbarPostServerItemToken");
 
     if (!rollbarKey) {
       console.log("Don't have rollbar so not logging error in prod?"); // eslint-disable-line no-console
@@ -155,7 +191,7 @@ class Macros extends commonMacros {
       return;
     }
 
-    const stack = (new Error()).stack;
+    const stack = new Error().stack;
 
     // The middle object can include any properties and values, much like amplitude.
     args.stack = stack;
@@ -189,19 +225,17 @@ class Macros extends commonMacros {
     }
   }
 
-
   // This is for programming errors. This will cause the program to exit anywhere.
   // This *should* never be called.
   static critical(...args: any) {
     if (Macros.TEST) {
-      console.error('macros.critical called'); // eslint-disable-line no-console
+      console.error("macros.critical called"); // eslint-disable-line no-console
       console.error(...args); // eslint-disable-line no-console
     } else {
       Macros.error(...args);
       process.exit(1);
     }
   }
-
 
   // Use this for stuff that is bad, and shouldn't happen, but isn't mission critical and can be ignored and the app will continue working
   // Will log something to rollbar and rollbar will send off an email
@@ -212,7 +246,6 @@ class Macros extends commonMacros {
       this.logRollbarError(args, false);
     }
   }
-
 
   // Use this for stuff that should never happen, but does not mean the program cannot continue.
   // This will continue running in dev, but will exit on CI
@@ -227,7 +260,7 @@ class Macros extends commonMacros {
       if (process.env.CI) {
         process.exit(1);
 
-      // If running on AWS, tell rollbar about the error so rollbar sends off an email.
+        // If running on AWS, tell rollbar about the error so rollbar sends off an email.
       } else {
         this.logRollbarError(args, true);
       }
@@ -244,13 +277,11 @@ class Macros extends commonMacros {
   }
 }
 
-
-Macros.verbose('Starting in verbose mode.');
-
+Macros.verbose("Starting in verbose mode.");
 
 async function handleUncaught(err: Error) {
   // Don't use the macros.log, because if that crashes it would run into an infinite loop
-  console.log('Error: An unhandledRejection occurred.'); // eslint-disable-line no-console
+  console.log("Error: An unhandledRejection occurred."); // eslint-disable-line no-console
   console.log(`Rejection Stack Trace: ${err.stack}`); // eslint-disable-line no-console
   Macros.error(err.stack);
 }
@@ -259,9 +290,8 @@ let addedRejectionHandler: boolean;
 // Sometimes it helps debugging to enable this test mode too.
 if (Macros.PROD && !addedRejectionHandler) {
   addedRejectionHandler = true;
-  process.on('unhandledRejection', handleUncaught);
-  process.on('uncaughtException', handleUncaught);
+  process.on("unhandledRejection", handleUncaught);
+  process.on("uncaughtException", handleUncaught);
 }
-
 
 export default Macros;
